pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "patchwork-diff"
version = "0.1.0"
description = "A semantic code diff tool powered by tree-sitter"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    { name = "Vedant", email = "your-email@example.com" }
]
dependencies = [
    "click",
    "rich",
    "textual",
    "gitpython",
    "tree-sitter",
    "tree-sitter-python",
    "tree-sitter-javascript",
]

[project.urls]
Homepage = "https://github.com/your-username/patchwork"

[project.scripts]
patchwork = "patchwork.cli:cli"

[tool.setuptools.packages.find]
include = ["patchwork*"]
